+++ src/client/pythonEnvironments/common/environmentIdentifier.ts
@@ -5,7 +5,7 @@ import { PythonEnvKind } from '../base/info';
 import { getPrioritizedEnvKinds } from '../base/info/envKind';
 import { isCondaEnvironment } from './environmentManagers/conda';
 import { isPipenvEnvironment } from '../discovery/locators/services/pipEnvHelper';
-import { isPoetryEnvironment } from '../discovery/locators/services/poetry';
+import { isPoetryEnvironment } from './environmentManagers/poetry';
 import { isPyenvEnvironment } from '../discovery/locators/services/pyenvLocator';
 import {
     isVenvEnvironment,
