+++ src/client/pythonEnvironments/discovery/locators/services/poetryLocator.ts
@@ -11,7 +11,7 @@ import { BasicEnvInfo, IPythonEnvsIterator } from '../../../base/locator';
 import { FSWatchingLocator } from '../../../base/locators/lowLevel/fsWatchingLocator';
 import { getInterpreterPathFromDir } from '../../../common/commonUtils';
 import { pathExists } from '../../../common/externalDependencies';
-import { isPoetryEnvironment, localPoetryEnvDirName, Poetry } from './poetry';
+import { isPoetryEnvironment, localPoetryEnvDirName, Poetry } from '../../../common/environmentManagers/poetry';
 import '../../../../common/extensions';
 import { asyncFilter } from '../../../../common/utils/arrayUtils';
